pyproject.toml
[project]
name = ""
version = "0.0.1"
description = ""
license = { text = "MIT" }
authors = [
    { name = "Doctor", email = "thirvondukr@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
    "uvloop>=0.17.0; sys_platform != 'win32'",
    "uvicorn>=0.23.2",
    "starlette>=0.27.0",
    "fastapi>=0.104.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "asyncpg>=0.28.0",
    "sqlalchemy>=2.0.22",
    "alembic>=1.12.1",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.10",
    "sentry-sdk>=1.32.0",
    "result>=0.13.1",
    "strawberry-graphql>=0.211.1",
    "aioinject>=0.15.0",
]
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.10.1",
    "ruff>=0.1.3",
    "mypy>=1.6.1",
    "radon>=6.0.1",
    "deptry>=0.12.0",
]
testing = [
    "pytest>=7.4.3",
    "coverage>=7.3.2",
    "httpx>=0.25.0",
    "anyio>=3.7.1",
    "sqlalchemy-pytest>=0.2.1",
    "asgi-lifespan>=2.1.0",
    "pytest-archon>=0.0.5",
    "typeguard>=4.1.5",
]

[tool.pdm.scripts]
lint = { composite = [
    "black .",
    "ruff . --fix",
    "black .",
    "deptry ./src",
    "mypy .",
] }
test = { composite = [
    "coverage run",
    "coverage report",
] }


[tool.pytest.ini_options]
pythonpath = "src"


[tool.coverage.run]
source = ["src"]
omit = ["src/alembic/*"]
command_line = "-m pytest -v --reuse-db"
concurrency = ["greenlet"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "class .*\\(.*\\bProtocol\\b.*\\):",
    "@(?:typing\\.)?overload",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
skip_covered = true
skip_empty = true
precision = 2
fail_under = 100


[tool.deptry]
[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn", "uvloop", "asyncpg", "orjson"]

[tool.deptry.package_module_name_map]
uvloop = "uvloop"

[tool.mypy]
plugins = [
    "pydantic.mypy",
    "strawberry.ext.mypy_plugin",
]
exclude = "(alembic)"

strict = true
follow_imports = "normal"
ignore_missing_imports = false

allow_redefinition = false
disallow_any_explicit = false
ignore_errors = false
local_partial_types = true
no_implicit_optional = true
strict_optional = true
warn_no_return = true
warn_return_any = false
warn_unreachable = true

pretty = true
show_column_numbers = true
show_error_codes = true

[tool.ruff]
src = ["src", "tests"]
fixable = [
    "F",
    "E",
    "W",
    "I",
    "COM",
    "UP",
    "RUF",
]
unfixable = [
    "F841", # Variable  is assigned to but never used
]
select = ["ALL"]
ignore = [
    "E501", # Line Length
    "ANN101",
    "ANN102",
    "D10", # Disable mandatory docstrings
    "D203", # one-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "PD", # pandas-vet
    "TCH001",
    "TCH002",
    "TCH003",
    "EXE",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
"src/app/alembic/*" = ["INP001"]
"src/app/alembic/versions/*" = ["D400", "D415"]
"src/app/db/models/*" = ["A003"]
"src/app/*/schema.py" = ["A003"]

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.mccabe]
max-complexity = 6

[tool.ruff.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.pep8-naming]
classmethod-decorators = ["classmethod"]
staticmethod-decorators = ["staticmethod"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"
